Crate partner_chains_smart_contracts_commands

This crate provides types and functions that can be used to create a CLI for managing the mainchain smart contracts relevant for a given partner chain instance.

Common Arguments

Most type commands (usualy ending in “Cmd”) take a CommonArguments struct as argument. It stores the information neccessary for connecting to the Ogmios server and retrying the operations like checking if a transaction is included in the blockchain.

Subcommands

Each subcommand has its own command type, which implements the [clap::Parser] trait. Each command type also has a execute method, which is used to execute the command.

Subcommands can execute transactions on the mainchain, query the mainchain and also provide other utilities for managing the smart contracts.

Result types

Most commands return a [SubCmdResult] type, which is a wrapper around a serde_json::Value. The returned value is printed to the ouptut at the end of the command execution.

Modules

assemble_tx

d_parameter

get_scripts

governance

governed_map

permissioned_candidates

register

reserve

sign_tx

Structs

CommonArguments

Common command arguments

Enums

SmartContractsCmd

Commands for managing the mainchain smart contracts